refactor(sql): extract the finding type system into sql/findings.py (#176)
`FindingKind`, `Finding`, `finding_at`, `suppression_code`, and
`suppression_hint` were the project-wide finding vocabulary living in
`sql/patterns.py`, a module named for one detector layer. Detectors across
nullability, uniqueness, snapshot, and flatten emit `FindingKind`s that
`patterns.py` never detects, so the type system reads as shared infrastructure
rather than a pattern.

Move those five symbols into `sql/findings.py`, the span-in-one-statement
counterpart to the declaration-level `check/findings.py`. `patterns.py` now
imports the finding types like every other detector module. The `dblect.sql`
re-exports stay identical, so consumers do not churn; the two tests that reached
into the private `dblect.sql.patterns` now import from the public surface.
